Seperated functions
- Created render.js
- Exported function from index.js to render.js
- Setup test cases
