+++ webpack.config.js
@@ -2,11 +2,6 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-  "jest": {
-    "moduleNameMapper": {
-      "\\.(css|scss)$": "<rootDir>/__mocks__/styleMock.js"
-    },
-  },
   mode: 'development',
   entry: './src/index.js',
   devtool: 'inline-source-map',
